chore: move quizSets to constants package
