+++ src/shared/utils/index.ts
@@ -1,2 +1,1 @@
 export { getRoute } from "./getRoute";
-export * from "./quizSets";
